+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Changelog
 This file is for listing all the changes to this project
 
-## v1.1.0 Release | Nov 8, 2023
+## v1.1.0 Release | Nov 11, 2023
 ### Adding
+- Adding Arabic Translation
 - Progression Display `You can now pin any challenges you want on the screen.`
+- Is Placed By Entity Condition Option `You can Enable/Disable the Is Placed By Entity Condition.`
+- Adjustable Mining Skill Card's Potion Points `You can change the potion points values in each tier.`
 - Adjustable Number of Challenges for each Tier Config `You can change the values on how many challenges should be given in each tier.`
   *But remember that you must have the exact number of challenges in the Datapack. Otherwise, it will make the game crash!*
 ### Fixed